Separate the http api with the plot api
