+++ ert_gui/tools/plot/plot_window.py
@@ -10,12 +10,13 @@ from ert_gui.plottery.plots.statistics import StatisticsPlot
 from ert_shared import ERT
 from ert_gui.ertwidgets import showWaitCursorWhileWaiting
 from ert_gui.plottery import PlotContext, PlotConfig
+from ert_shared.feature_toggling import FeatureToggling
 
 from ert_gui.tools.plot import DataTypeKeysWidget, CaseSelectionWidget, PlotWidget
 from ert_gui.tools.plot.customize import PlotCustomizer
 
 from ert_gui.tools.plot.plot_api import PlotApi
-from ert_shared.storage.storage_api import StorageApi
+from ert_gui.tools.plot.storage_client import StorageClient
 import pandas as pd
 
 CROSS_CASE_STATISTICS = "Cross Case Statistics"
@@ -31,8 +32,10 @@ class PlotWindow(QMainWindow):
     def __init__(self, config_file, parent):
         QMainWindow.__init__(self, parent)
 
-        #self._api = PlotApi(ERT.enkf_facade)
-        self._api = StorageApi()
+        if FeatureToggling.is_enabled("new-storage"):
+            self._api = StorageClient()
+        else:
+            self._api = PlotApi(ERT.enkf_facade)
 
         self.setMinimumWidth(850)
         self.setMinimumHeight(650)
