+++ tests/gui/tools/plot/test_plot_storage_api.py
@@ -3,18 +3,18 @@ from unittest import TestCase
 import pandas as pd
 import pprint
 
-from ert_shared.storage.storage_api import StorageApi
+from ert_gui.tools.plot.storage_client import StorageClient
 
 from tests.storage import populated_db, db_session, engine, tables
 from io import StringIO
 
 def test_all_keys(populated_db):
-    api = StorageApi(populated_db)
+    api = StorageClient(populated_db)
     names = set([key['key'] for key in api.all_data_type_keys()])
     assert names == set(["response_one", "response_two", "A", "B"])
 
 def test_observation_values(populated_db):
-    api = StorageApi(populated_db)
+    api = StorageClient(populated_db)
     result = api.observations_for_obs_keys(case="ensemble_name", obs_keys=["observation_one"])
     idx = pd.MultiIndex.from_arrays([[2, 3], [0, 3]],
                                     names=['data_index', 'key_index'])
@@ -26,14 +26,14 @@ def test_parameter_values():
     pass
 
 def test_response_values(populated_db):
-    api = StorageApi(populated_db)
+    api = StorageClient(populated_db)
     result = api.data_for_key(case="ensemble_name", key="response_one")
     idx = pd.MultiIndex.from_arrays([["response_one", "response_one"], ['0', '1']], names=["keys", "indexes"])
     expected = pd.DataFrame([[11.1, 11.1], [11.2, 11.2]], index=idx).T
     assert result.equals(expected)
 
 def test_stream_param(populated_db):
-    api = StorageApi(populated_db)
+    api = StorageClient(populated_db)
     for msg in api.get_param_data(ensembles = ["ensemble_name"]):
         print(str(msg))
 
@@ -44,20 +44,3 @@ def test_stream_param(populated_db):
     print(csv)
     df = pd.read_csv(StringIO(csv), index_col=[0], header=[0, 1])
     print(df)
-
-def test_ensemble_schema(populated_db):
-    api = StorageApi(populated_db)
-
-
-    schema = api.ensemble_schema("ensemble_name")
-    pp = pprint.PrettyPrinter(indent=2)
-    pp.pprint(schema)
-
-def test_ensembles(populated_db):
-    api = StorageApi(populated_db)
-
-
-    ensembles = api.ensembles()
-    pp = pprint.PrettyPrinter(indent=2)
-    pp.pprint(ensembles)
-
